requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
streamlit>=1.48.0
requests>=2.32.0
python-dotenv>=1.0.1
pydantic>=2.11.0
pydantic-settings>=2.10.0
langchain>=0.3.27
langgraph>=0.6.0
langchain-groq>=0.3.6
langchain-community>=0.3.27
wikipedia>=1.4.0
duckduckgo-search>=6.3.0
structlog>=25.4.0
httpx>=0.28.0
pytest>=8.3.0
pytest-asyncio>=1.1.0
ddgs
